refactor(python): split mpp server charge into decode and verify modules
The server charge module had grown to 1612 lines, mixing three concerns:
pure transaction decoders, the fee-payer cosign plus no-leftovers
instruction allowlist, and the Mpp orchestration class. Extract the
decoders into _tx_decode and the cosign/allowlist into _verify, leaving
charge.py at 612 lines holding only the Mpp flow. The moved helpers are
re-exported from charge so the existing import surface and the tests that
reach into the private helpers stay unchanged. No behavior change.
